mrail: Store sequence number with peer addresses
Messages must be processed at the receiver in the same order that
they were sent in order for tag matching to work correctly.

For the receiver to do this, have the sender set a sequence
number on each message.  The sequence number is kept per
peer.

Define a new mrail_peer_addr.  This will store each fi_addr
returned by each rail's AV, along with per peer data, such
as a seq_no.  Update message and rma operations to use this
new structure.  This includes updating the mrail_req struct.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>
Signed-off-by: yohann <yohann.burette@intel.com>