prov/mail: Use FI_AV_TABLE from underlying providers
The mrail provider basically converts the underlying provider
AV into AV table.  By moving this down, we can avoid the extra storage
in the mrail provider when the underlying provider is already using
AV table.  It also means that the fi_addr_t passed in by the app
can be handed directly to the underlying provider, which avoids a
conversion.

As an additional benefit, the fi_addr_t passed up from the lower
provider maps directly to that handed to the app.  This avoids
mrail from needing a reverse lookup map/hash table for each rail.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>